Topo.py
Select Model:選擇高模
Import Modelg:執行生成 Diffuse+Normal+Low_obj.fbx+Low_obj.blend


This is a Python script that uses the Blender Python API to import an FBX model, apply modifications to it, and add materials to it. The script uses the tkinter module to create a simple GUI with two buttons that allow the user to select an FBX file and import it into Blender.
The FbxImporter class is defined with an init method that creates the GUI and sets the default file paths. The select_model method is called when the "Select Model" button is clicked and uses the filedialog module from tkinter to open a file dialog window for the user to select an FBX file. The import_model method is called when the "Import Model" button is clicked and applies modifications to the imported model.
The modifications include clearing the scene, importing the model, selecting and making it active, applying WeightedNormal, setting Auto Smooth, renaming the object, duplicating it, applying a Remesh modifier, and performing a Smart UV Project. Finally, the script adds a material to the low-poly version of the model and sets up nodes for the Normal map.
This script is useful for importing and modifying FBX models in Blender, especially for creating low-poly models from high-poly ones. It also demonstrates the use of the Blender Python API and the integration of Python with a GUI library like tkinter.